refactor: move task-related packages under task module
Organize imports and package paths to group all task-related code
under the "task" module. This improves project structure and
modularity by clearly separating task domain, application,
infrastructure, handlers, and DTOs from other parts of the codebase.
