+++ task/infrastructure/task_repository.go
@@ -2,7 +2,7 @@ package infrastructure
 
 import (
 	"context"
-	"github.com/ltphat2204/domain-driven-golang/domain"
+	"github.com/ltphat2204/domain-driven-golang/task/domain"
 	"gorm.io/gorm"
 )
